+++ CHANGELOG.md
@@ -1,11 +1,16 @@
-##[1.0.4+1] - 20200119
+## [1.0.5] - 20200202
+* Feat: Added new elevationColor allowing change of shadow color
+* Improve: Use boxShadow avoid flutter Material shadow issue 
+* Fix: Horizontal shadow cannot be shown due to current flutter version
+
+## [1.0.4+1] - 20200119
 * Fix: README.md background color default color as white
 
-##[1.0.4] - 20200119
+## [1.0.4] - 20200119
 * Fix: left hand side column show incorrect color
 * Doc: updated example of doing sorting
 
-##[1.0.3] - 20200119
+## [1.0.3] - 20200119
 * Improve: use Selector instead of Consumer 
 
 ## [1.0.2] - 20200117